Illustration of old Fortran (1977)
In that old version of Fortran, you need to start your program at
column 6 and any one line cannot be longer than 72
characters (constraints imposed by the punchcards).

The program has a bug that is not detected by the compiler (on my
computer it shows *40091396*. I guess it was supposed to show integers
from 1 to 20?